requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# Flask Extensions
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# IA and OpenAI
openai==1.3.7
requests==2.31.0

# Export and Document Generation
reportlab==4.0.7
markdown==3.5.1
python-markdown-math==0.8

# Configuration and Environment
python-dotenv==1.0.0
click==8.1.7

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
coverage==7.3.2

# Security and Validation
WTForms==3.1.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pathlib2==2.3.7 